rename LazyFocusedPolling to useLazyPolling
